Extract class for verification of script templates
Extract a class from groovy's VerifyScriptTemplates class to match the
class used by other scripting (beta) add-ons (e.g. Zest, Jython, JRuby),
to reduce the changes needed when merging the branches.
